blanc-flask/requirements.txt
alembic==1.4.2
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
firebase==3.0.1
Flask==1.1.2
glog==0.3.1
google-api-core==1.20.0
google-auth==1.17.2
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.1
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
memoization==0.3.1
oauth2client==4.1.3
protobuf==3.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglib==0.1
python-dateutil==2.8.1
python-editor==1.0.4
python-gflags==3.1.2
pytz==2020.1
requests
rsa==4.1
six==1.15.0
sqlparse==0.3.1
urllib3
Werkzeug==1.0.1
zipp==3.1.0
pendulum~=2.1.0
mongoengine~=0.20.0
pymongo==3.5.1
flask_mongoengine
firebase_admin==4.4.0
python-dotenv
mock
mongomock
google-api-python-client
gunicorn
inapppy